Add the 'get_xml_tag' function to get the XML data from tags regardless of the namespace
This function implements the retrieval of the data in the XML object
regardless of the namespace where this information is. Depending upon
the AFIS vendor, the namespaces can be "biom:" or "ansi-nist:"; if we
fix the namespace to "biom:" like it was the case, all the NIST files
with other namespaces can not be parsed.