Add the possible Ridge Count field in the minutiae tag
Reading the NIST standard, the ridge count can be present in the
minutiae field. By simply splitting the field with the normal separator,
it's not possible, without counting the number of fields retrieved, to
know if the RC field is present or not, hence not possible to use the
variable unpacking.