Skip empty fields on load - Missing ending 1C
This is done because of a US NIST file that was off by one byte at the
end of the file. This can append if the last byte is not a "1C". If this
skip produce a valid NIST object (i.e there is not fields after), it's
probable that the entire NIST file has been processed correctly.

If this error is present in the middle of the file, the processing of
the next record should fail and crash completely the loading of the file
(hopefully...).